requirements.txt
click==8.1.7
jsonschema[format]==4.21.1
pyyaml==6.0.1
